refactor: introduce `eventstreaming` package in `messaging` module
