+++ eventsourcing/src/test/java/org/axonframework/eventsourcing/eventstore/FilteringDomainEventStreamTest.java
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-package org.axonframework.eventsourcing;
+package org.axonframework.eventsourcing.eventstore;
 
 import org.axonframework.eventhandling.DomainEventMessage;
 import org.axonframework.eventhandling.GenericDomainEventMessage;
-import org.axonframework.eventsourcing.eventstore.DomainEventStream;
-import org.axonframework.eventsourcing.eventstore.FilteringDomainEventStream;
 import org.axonframework.messaging.MessageType;
 import org.junit.jupiter.api.*;
